tests/wpt/web-platform-tests/workers/interfaces/WorkerGlobalScope/close/setTimeout.html

<!--
function x() {
  postMessage(1);
  throw new Error();
}
setTimeout(x, 0);
close();
setTimeout(x, 0);
/*
-->
<!doctype html>
<title>close() and setTimeout</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
setup({allow_uncaught_exception: true});

async_test(function() {
  var gotMessage = false;
  var gotError = false;
  var worker = new Worker('#');
  worker.onmessage = function(e) {
    gotMessage = true;
  };
  worker.onerror = function(e) {
    gotError = true;
  };
  setTimeout(this.step_func(function() {
    assert_false(gotMessage, 'got message');
    assert_true(gotError, 'got error');
    this.done();
  }), 500);
});
</script>
<!--
*/
//-->
